+++ Changelog.md
@@ -7,3 +7,8 @@
 - Added optional save before application
 - More bug fixes to prevent multiple application(check for auras before running workflow on token create)
 - Auras no-longer apply when a token is at 0 hp (optional toggle in the settings)
+
+## 0.1.43
+- Removed euclidean distance option, auras now only use grid measurment distance
+- Added Korean Localization
+- Fixed a localization issue